LiveOrg: Support finding the daystate id and rule set for a given date
A lot of rule-specific stuff was moved to a new file, some of it is
being tested, most is not. The daystate of a particular day is cached
up until a new fixed state is pushed, causing the cache to invalidate
from the date of the fix onward. Rule sets have been separated into
one list containing only reoccurring rules while non-reoccurring
patterns are put into the single-use bucket / sorted set where the
score determines the timestamp the rule set applies to. This makes it
really easy to get the rule sets that were made for a particular day.